backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.22.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0

# LangChain and LLM
langchain==1.2.0
langchain-deepseek==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.6

# Data Processing
pandas>=2.0.0
numpy>=1.26.0
pydantic==2.10.0
pydantic-settings==2.12.0
annotated-types>=0.6.0

# External Data
yfinance>=1.0.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0

# Scheduler
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
loguru>=0.7.0
httpx>=0.26.0
